desktop-manager-api/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[project]
name = "desktop_manager"
version = "0.1.0"
description = "Desktop Manager API"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "flask-cors",
    "gunicorn",
    "mysqlclient",
    "sqlalchemy",
    "requests",
    "pydantic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "kubernetes",
    "werkzeug",
    "PyJWT",
    # Security dependencies
    "argon2-cffi",
    "bleach",
    "pydantic[email]",
    "flask-limiter",
    "flask-talisman",
    "flask-seasurf",
    "email-validator"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["desktop_manager*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
